requirements.txt
openai
flask
flask[async]
python-telegram-bot